# Conflicts:
#	.idea/workspace.xml

+++ src/main/java/com/example/letsGo/service/RecordService.java
@@ -1,14 +1,15 @@
 package com.example.letsGo.service;
 
+import com.example.letsGo.domain.record.Record;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public interface RecordService {
-    Record getRecord(int recordId);
+    Record getRecord(Long recordId);
     List<Record> getAllRecords(int userId);
     Record addRecord(Record record);
-    Record updateRecord(int recordId, Record record);
-    void deleteRecord(int recordId);
+    Record updateRecord(Long recordId, Record record);
+    void deleteRecord(Long recordId);
 }
